dumb but absolutely the best thing I ever did
All the script does is open a window that lets a user click on a job file and have the file path from a certain directory copied to the clipboard.
This is absolutely a hack, but we used the file path as the job title stored in our job databases. This script eliminated the chance that the path would be entered incorrectly and was the foundation of every automation we enacted after that.